requirements.txt
numpy>=1.20.0
pillow>=9.0.0
streamlit>=1.20.0
scikit-image>=0.19.0
scikit-learn>=1.0.0
joblib>=1.1.0
huggingface_hub>=0.13.0
torch>=1.12.0
torchvision>=0.13.0
matplotlib>=3.5.0
tqdm>=4.62.0 